Add files via upload
Suppose that the row of Gryffindor's potions contains a sequence of n potions whose n-th potion strength is equal to a i. Also suppose Slytherin potions include a sequence Let m be a set of potions whose i-th potion strength is b i Is. Since Professor Snape is a just man, he wants to choose a subsequence from the potions of each group, so that this subsequence of potions has potions of equal strength.